Browse available RBA data series

Use these functions to find the table number or series ID of the data you're interested in.

browse_rba_series(search_string = "")

browse_rba_tables(search_string = "", refresh = FALSE)

Arguments

search_string
Word or phrase to search for, such as "gold" or "commodity" or "labour". If left as "", the function will return all series.

refresh
logical; FALSE by default. When FALSE, internal data is used. When TRUE, the RBA website is re-scraped to obtain current information about available tables This can take a few seconds.

Value

A data.frame (tbl_df) containing RBA data series/tables that match the search_string. Where no search_string is supplied, the data.frame will contain information about all RBA series/tables.

The data.frame returned by browse_rba_tables() includes a column called readable. This column takes the value TRUE if the table is able to be read by read_rba() and FALSE if it cannot be read.

Examples

# Find series that contain 'unemployment'
browse_rba_series("unemployment")

#> # A tibble: 2 × 8
#>   table_no series            series_id series_type table_title cur_hist description
#>   <chr>    <chr>             <chr>     <chr>       <chr>       <chr>    <chr>      
#> 1 H5       Unemployment      GLFSUPSA  Seasonally… H5 Labour … current  Unemployed…
#> 2 H5       Unemployment rate GLFSURSA  Seasonally… H5 Labour … current  Unemployed…
#> # … with 1 more variable: frequency <chr>

# Or all labour-related series
browse_rba_series("labour")

#> # A tibble: 23 × 8
#>    table_no series    series_id series_type table_title  cur_hist description   
#>    <chr>    <chr>     <chr>     <chr>       <chr>        <chr>    <chr>         
#>  1 H4       Non-farm… GNFPROSQI Seasonally… H4 Labour C… current  Non-farm labo…
#>  2 H4       Quarterl… GWPIPRIQP Seasonally… H4 Labour C… current  Wage price in…
#>  3 H4       Quarterl… GWPIPUBQP Seasonally… H4 Labour C… current  Wage price in…
#>  4 H4       Quarterl… GWPIQP    Seasonally… H4 Labour C… current  Wage price in…
#>  5 H4       Year-end… GNFAEYP   Seasonally… H4 Labour C… current  Non-farm aver…
#>  6 H4       Year-end… GNFPROSQP Seasonally… H4 Labour C… current  Non-farm labo…
#>  7 H4       Year-end… GNFULCYP  Seasonally… H4 Labour C… current  Non-farm nomi…
#>  8 H4       Year-end… GWPIPRIYP Seasonally… H4 Labour C… current  Wage price in…
#>  9 H4       Year-end… GWPIPUBYP Seasonally… H4 Labour C… current  Wage price in…
#> 10 H4       Year-end… GWPIYP    Seasonally… H4 Labour C… current  Wage price in…
#> # … with 13 more rows, and 1 more variable: frequency <chr>

# Or those related to commodities
browse_rba_series("commodities")

#> # A tibble: 9 × 8
#>   table_no series     series_id series_type table_title cur_hist description    
#>   <chr>    <chr>      <chr>     <chr>       <chr>       <chr>    <chr>          
#> 1 I2       Bulk comm… GRCPBCAD  Original    I2 Commodi… current  Index of commo…
#> 2 I2       Bulk comm… GRCPBCSDR Original    I2 Commodi… current  Index of commo…
#> 3 I2       Bulk comm… GRCPBCUSD Original    I2 Commodi… current  Index of commo…
#> 4 I2       Bulk comm… GRCPBCSAD Original    I2 Commodi… current  Index of commo…
#> 5 I2       Bulk comm… GRCPBCSS… Original    I2 Commodi… current  Index of commo…
#> 6 I2       Bulk comm… GRCPBCSU… Original    I2 Commodi… current  Index of commo…
#> 7 I2       Commodity… GRCPAISAD Original    I2 Commodi… current  Index of commo…
#> 8 I2       Commodity… GRCPAISS… Original    I2 Commodi… current  Index of commo…
#> 9 I2       Commodity… GRCPAISU… Original    I2 Commodi… current  Index of commo…
#> # … with 1 more variable: frequency <chr>

# Or all series
browse_rba_series()

#> # A tibble: 4,412 × 8
#>    table_no series   series_id series_type table_title    cur_hist description  
#>    <chr>    <chr>    <chr>     <chr>       <chr>          <chr>    <chr>        
#>  1 A1       Austral… ARBAAASTW Original    A1 Reserve Ba… current  Australian d…
#>  2 A1       Capital… ARBALCRFW Original    A1 Reserve Ba… current  Capital and …
#>  3 A1       Deposit… ARBALDEPW Original    A1 Reserve Ba… current  Deposits (ex…
#>  4 A1       Exchang… ARBALESBW Original    A1 Reserve Ba… current  Exchange set…
#>  5 A1       Gold an… ARBAAGFXW Original    A1 Reserve Ba… current  Gold and for…
#>  6 A1       Notes o… ARBALNOIW Original    A1 Reserve Ba… current  Notes on iss…
#>  7 A1       Other a… ARBAAOAW  Original    A1 Reserve Ba… current  Other assets…
#>  8 A1       Other l… ARBALOLW  Original    A1 Reserve Ba… current  Other liabil…
#>  9 A1       Total a… ARBAATAW  Original    A1 Reserve Ba… current  Total RBA as…
#> 10 A1       Total l… ARBALTLW  Original    A1 Reserve Ba… current  Total RBA li…
#> # … with 4,402 more rows, and 1 more variable: frequency <chr>

# Or just look for tables that contain the word 'labour'
browse_rba_tables("labour")

#> # A tibble: 2 × 5
#>   title                         no    url              current_or_hist… readable
#>   <chr>                         <chr> <chr>            <chr>            <lgl>   
#> 1 Labour Costs and Productivity H4    https://www.rba… current          TRUE    
#> 2 Labour Force                  H5    https://www.rba… current          TRUE    

# Or all tables
browse_rba_tables()

#> # A tibble: 123 × 5
#>    title                   no    url                  current_or_histo… readable
#>    <chr>                   <chr> <chr>                <chr>             <lgl>   
#>  1 Liabilities and Assets… A1    https://www.rba.gov… current           TRUE    
#>  2 Liabilities and Assets… A1.1  https://www.rba.gov… current           TRUE    
#>  3 Monetary Policy Changes A2    https://www.rba.gov… current           TRUE    
#>  4 Monetary Policy Operat… A3    https://www.rba.gov… current           TRUE    
#>  5 Holdings of Australian… A3.1  https://www.rba.gov… current           TRUE    
#>  6 Securities Lending Rep… A3.2  https://www.rba.gov… current           TRUE    
#>  7 Foreign Exchange Trans… A4    https://www.rba.gov… current           TRUE    
#>  8 Daily Foreign Exchange… A5    https://www.rba.gov… current           TRUE    
#>  9 Banknotes on Issue by … A6    https://www.rba.gov… current           TRUE    
#> 10 Detected Australian Co… A7    https://www.rba.gov… current           TRUE    
#> # … with 113 more rows

# To re-scrape the RBA website to ensure you have up-to-date information
# about available tables:
if (FALSE) {
browse_rba_tables(refresh = TRUE)
}