Convenience function to get the RBA's cash rate.

Convenience function to get the RBA's cash rate.

read_cashrate(type = c("target", "interbank", "both"))

rba_cashrate(...)

Arguments

type

One of 'target' (the default), 'interbank', or 'both'.

'target'

The RBA's cash rate target.

'interbank'

The interbank overnight cash rate.

'both'

Both the cash rate target and interbank overnight cash rate.

'target' fetches the RBA cash rate target. 'interbank'

...

arguments passed to read_cashrate()

Value

A tbl_df with two columns: date and cash_rate.

Details

Note that in the very early 1990s, the cash rate target was expressed as a range (eg. "17% to 17.5%"). Where this is the case, the value returned here (and in read_rba()) is the mid-point of this range.

If type = 'both', note that the returned tbl is tidy/long.

rba_cashrate() is a wrapper around read_cashrate().