Browse available RBA data series

Use these functions to find the table number or series ID of the data you're interested in.

browse_rba_series(search_string = "")

browse_rba_tables(search_string = "", refresh = FALSE)

Arguments

search_string

Word or phrase to search for, such as "gold" or "commodity" or "labour". If left as "", the function will return all series.

refresh

logical; FALSE by default. When FALSE, internal data is used. When TRUE, the RBA website is re-scraped to obtain current information about available tables This can take a few seconds.

Value

A data.frame (tbl_df) containing RBA data series/tables that match the search_string. Where no search_string is supplied, the data.frame will contain information about all RBA series/tables.

The data.frame returned by browse_rba_tables() includes a column called readable. This column takes the value TRUE if the table is able to be read by read_rba() and FALSE if it cannot be read.

Examples

# Find series that contain 'unemployment'
browse_rba_series("unemployment")
#> # A tibble: 2 × 8
#>   table_no series            series_id series_…¹ table…² cur_h…³ descr…⁴ frequ…⁵
#>   <chr>    <chr>             <chr>     <chr>     <chr>   <chr>   <chr>   <chr>  
#> 1 H5       Unemployment      GLFSUPSA  Seasonal… H5 Lab… current Unempl… Monthly
#> 2 H5       Unemployment rate GLFSURSA  Seasonal… H5 Lab… current Unempl… Monthly
#> # … with abbreviated variable names ¹​series_type, ²​table_title, ³​cur_hist,
#> #   ⁴​description, ⁵​frequency

# Or all labour-related series
browse_rba_series("labour")
#> # A tibble: 23 × 8
#>    table_no series               serie…¹ serie…² table…³ cur_h…⁴ descr…⁵ frequ…⁶
#>    <chr>    <chr>                <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 H4       Non-farm labour pro… GNFPRO… Season… H4 Lab… current Non-fa… Quarte…
#>  2 H4       Quarterly private s… GWPIPR… Season… H4 Lab… current Wage p… Quarte…
#>  3 H4       Quarterly public se… GWPIPU… Season… H4 Lab… current Wage p… Quarte…
#>  4 H4       Quarterly wage grow… GWPIQP  Season… H4 Lab… current Wage p… Quarte…
#>  5 H4       Year-ended non-farm… GNFAEYP Season… H4 Lab… current Non-fa… Quarte…
#>  6 H4       Year-ended non-farm… GNFPRO… Season… H4 Lab… current Non-fa… Quarte…
#>  7 H4       Year-ended non-farm… GNFULC… Season… H4 Lab… current Non-fa… Quarte…
#>  8 H4       Year-ended private … GWPIPR… Season… H4 Lab… current Wage p… Quarte…
#>  9 H4       Year-ended public s… GWPIPU… Season… H4 Lab… current Wage p… Quarte…
#> 10 H4       Year-ended wage gro… GWPIYP  Season… H4 Lab… current Wage p… Quarte…
#> # … with 13 more rows, and abbreviated variable names ¹​series_id, ²​series_type,
#> #   ³​table_title, ⁴​cur_hist, ⁵​description, ⁶​frequency

# Or those related to commodities
browse_rba_series("commodities")
#> # A tibble: 9 × 8
#>   table_no series                serie…¹ serie…² table…³ cur_h…⁴ descr…⁵ frequ…⁶
#>   <chr>    <chr>                 <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 I2       Bulk commodities pri… GRCPBC… Origin… I2 Com… current Index … Monthly
#> 2 I2       Bulk commodities pri… GRCPBC… Origin… I2 Com… current Index … Monthly
#> 3 I2       Bulk commodities pri… GRCPBC… Origin… I2 Com… current Index … Monthly
#> 4 I2       Bulk commodities spo… GRCPBC… Origin… I2 Com… current Index … Monthly
#> 5 I2       Bulk commodities spo… GRCPBC… Origin… I2 Com… current Index … Monthly
#> 6 I2       Bulk commodities spo… GRCPBC… Origin… I2 Com… current Index … Monthly
#> 7 I2       Commodity prices (wi… GRCPAI… Origin… I2 Com… current Index … Monthly
#> 8 I2       Commodity prices (wi… GRCPAI… Origin… I2 Com… current Index … Monthly
#> 9 I2       Commodity prices (wi… GRCPAI… Origin… I2 Com… current Index … Monthly
#> # … with abbreviated variable names ¹​series_id, ²​series_type, ³​table_title,
#> #   ⁴​cur_hist, ⁵​description, ⁶​frequency

# Or all series
browse_rba_series()
#> # A tibble: 4,478 × 8
#>    table_no series               serie…¹ serie…² table…³ cur_h…⁴ descr…⁵ frequ…⁶
#>    <chr>    <chr>                <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 A1       Australian dollar i… ARBAAA… Origin… A1 Res… current Austra… Weekly 
#>  2 A1       Capital and Reserve… ARBALC… Origin… A1 Res… current Capita… Weekly 
#>  3 A1       Deposits (excluding… ARBALD… Origin… A1 Res… current Deposi… Weekly 
#>  4 A1       Exchange settlement… ARBALE… Origin… A1 Res… current Exchan… Weekly 
#>  5 A1       Gold and foreign ex… ARBAAG… Origin… A1 Res… current Gold a… Weekly 
#>  6 A1       Notes on issue       ARBALN… Origin… A1 Res… current Notes … Weekly 
#>  7 A1       Other assets (inclu… ARBAAO… Origin… A1 Res… current Other … Weekly 
#>  8 A1       Other liabilities    ARBALO… Origin… A1 Res… current Other … Weekly 
#>  9 A1       Total assets         ARBAAT… Origin… A1 Res… current Total … Weekly 
#> 10 A1       Total liabilities    ARBALT… Origin… A1 Res… current Total … Weekly 
#> # … with 4,468 more rows, and abbreviated variable names ¹​series_id,
#> #   ²​series_type, ³​table_title, ⁴​cur_hist, ⁵​description, ⁶​frequency

# Or just look for tables that contain the word 'labour'
browse_rba_tables("labour")
#> # A tibble: 2 × 5
#>   title                         no    url                        curre…¹ reada…²
#>   <chr>                         <chr> <chr>                      <chr>   <lgl>  
#> 1 Labour Costs and Productivity H4    https://www.rba.gov.au/st… current TRUE   
#> 2 Labour Force                  H5    https://www.rba.gov.au/st… current TRUE   
#> # … with abbreviated variable names ¹​current_or_historical, ²​readable

# Or all tables
browse_rba_tables()
#> # A tibble: 123 × 5
#>    title                                             no    url   curre…¹ reada…²
#>    <chr>                                             <chr> <chr> <chr>   <lgl>  
#>  1 Liabilities and Assets – Summary                  A1    http… current TRUE   
#>  2 Liabilities and Assets – Detailed                 A1.1  http… current TRUE   
#>  3 Monetary Policy Changes                           A2    http… current TRUE   
#>  4 Monetary Policy Operations – Current              A3    http… current TRUE   
#>  5 Holdings of Australian Government Securities and… A3.1  http… current TRUE   
#>  6 Securities Lending Repurchase and Switch Transac… A3.2  http… current TRUE   
#>  7 Foreign Exchange Transactions and Holdings of Of… A4    http… current TRUE   
#>  8 Daily Foreign Exchange Market Intervention Trans… A5    http… current TRUE   
#>  9 Banknotes on Issue by Denomination                A6    http… current TRUE   
#> 10 Detected Australian Counterfeits by Denomination  A7    http… current TRUE   
#> # … with 113 more rows, and abbreviated variable names ¹​current_or_historical,
#> #   ²​readable

# To re-scrape the RBA website to ensure you have up-to-date information
# about available tables:
if (FALSE) {
browse_rba_tables(refresh = TRUE)
}