Browse available RBA data series

Browse available RBA data series

browse_rba_series(search_string = "")

browse_rba_tables(search_string = "")

Arguments

search_string
Word or phrase to search for, such as "gold" or "commodity" or "labour". If left as "", the function will return all series.

Value

A data.frame (tbl_df) containing RBA data series/tables that match the search_string.

Examples

# Find series that contain 'unemployment'
browse_rba_series("unemployment")

#> # A tibble: 2 x 7
#>   table_no series   series_id series_type   table_title cur_hist description    
#>   <chr>    <chr>    <chr>     <chr>         <chr>       <chr>    <chr>          
#> 1 H5       Unemplo… GLFSUPSA  Seasonally a… H5 Labour … current  Unemployed per…
#> 2 H5       Unemplo… GLFSURSA  Seasonally a… H5 Labour … current  Unemployed per…

# Or all labour-related series
browse_rba_series("labour")

#> # A tibble: 23 x 7
#>    table_no series    series_id series_type table_title  cur_hist description   
#>    <chr>    <chr>     <chr>     <chr>       <chr>        <chr>    <chr>         
#>  1 H4       Non-farm… GNFPROSQI Seasonally… H4 Labour C… current  Non-farm labo…
#>  2 H4       Quarterl… GWPIPRIQP Seasonally… H4 Labour C… current  Wage price in…
#>  3 H4       Quarterl… GWPIPUBQP Seasonally… H4 Labour C… current  Wage price in…
#>  4 H4       Quarterl… GWPIQP    Seasonally… H4 Labour C… current  Wage price in…
#>  5 H4       Year-end… GNFAEYP   Seasonally… H4 Labour C… current  Non-farm aver…
#>  6 H4       Year-end… GNFPROSQP Seasonally… H4 Labour C… current  Non-farm labo…
#>  7 H4       Year-end… GNFULCYP  Seasonally… H4 Labour C… current  Non-farm nomi…
#>  8 H4       Year-end… GWPIPRIYP Seasonally… H4 Labour C… current  Wage price in…
#>  9 H4       Year-end… GWPIPUBYP Seasonally… H4 Labour C… current  Wage price in…
#> 10 H4       Year-end… GWPIYP    Seasonally… H4 Labour C… current  Wage price in…
#> # … with 13 more rows

# Or those related to commodities
browse_rba_series("commodities")

#> # A tibble: 9 x 7
#>   table_no series     series_id series_type table_title cur_hist description    
#>   <chr>    <chr>      <chr>     <chr>       <chr>       <chr>    <chr>          
#> 1 I2       Bulk comm… GRCPBCAD  Original    I2 Commodi… current  Index of commo…
#> 2 I2       Bulk comm… GRCPBCSDR Original    I2 Commodi… current  Index of commo…
#> 3 I2       Bulk comm… GRCPBCUSD Original    I2 Commodi… current  Index of commo…
#> 4 I2       Bulk comm… GRCPBCSAD Original    I2 Commodi… current  Index of commo…
#> 5 I2       Bulk comm… GRCPBCSS… Original    I2 Commodi… current  Index of commo…
#> 6 I2       Bulk comm… GRCPBCSU… Original    I2 Commodi… current  Index of commo…
#> 7 I2       Commodity… GRCPAISAD Original    I2 Commodi… current  Index of commo…
#> 8 I2       Commodity… GRCPAISS… Original    I2 Commodi… current  Index of commo…
#> 9 I2       Commodity… GRCPAISU… Original    I2 Commodi… current  Index of commo…

# Or all series
browse_rba_series()

#> # A tibble: 4,136 x 7
#>    table_no series   series_id series_type table_title    cur_hist description  
#>    <chr>    <chr>    <chr>     <chr>       <chr>          <chr>    <chr>        
#>  1 A1       Austral… ARBAAASTW Original    A1 Reserve Ba… current  Australian d…
#>  2 A1       Capital… ARBALCRFW Original    A1 Reserve Ba… current  Capital and …
#>  3 A1       Deposit… ARBALDEPW Original    A1 Reserve Ba… current  Deposits (ex…
#>  4 A1       Exchang… ARBALESBW Original    A1 Reserve Ba… current  Exchange set…
#>  5 A1       Gold an… ARBAAGFXW Original    A1 Reserve Ba… current  Gold and for…
#>  6 A1       Notes o… ARBALNOIW Original    A1 Reserve Ba… current  Notes on iss…
#>  7 A1       Other a… ARBAAOAW  Original    A1 Reserve Ba… current  Other assets…
#>  8 A1       Other l… ARBALOLW  Original    A1 Reserve Ba… current  Other liabil…
#>  9 A1       Total a… ARBAATAW  Original    A1 Reserve Ba… current  Total RBA as…
#> 10 A1       Total l… ARBALTLW  Original    A1 Reserve Ba… current  Total RBA li…
#> # … with 4,126 more rows

# Or just look for tables that contain the word 'labour'
browse_rba_tables("labour")

#> # A tibble: 2 x 4
#>   title               no    url                               current_or_histor…
#>   <chr>               <chr> <chr>                             <chr>             
#> 1 Labour Costs and P… H4    https://rba.gov.au/statistics/ta… current           
#> 2 Labour Force        H5    https://rba.gov.au/statistics/ta… current           

# Or all tables
browse_rba_tables()

#> # A tibble: 123 x 4
#>    title                       no    url                       current_or_histo…
#>    <chr>                       <chr> <chr>                     <chr>            
#>  1 Liabilities and Assets – S… A1    https://rba.gov.au/stati… current          
#>  2 Liabilities and Assets – D… A1.1  https://rba.gov.au/stati… current          
#>  3 Monetary Policy Changes     A2    https://rba.gov.au/stati… current          
#>  4 Monetary Policy Operations… A3    https://rba.gov.au/stati… current          
#>  5 Holdings of Australian Gov… A3.1  https://rba.gov.au/stati… current          
#>  6 Foreign Exchange Transacti… A4    https://rba.gov.au/stati… current          
#>  7 Daily Foreign Exchange Mar… A5    https://rba.gov.au/stati… current          
#>  8 Banknotes on Issue by Deno… A6    https://rba.gov.au/stati… current          
#>  9 Detected Australian Counte… A7    https://rba.gov.au/stati… current          
#> 10 Assets of Financial Instit… B1    https://rba.gov.au/stati… current          
#> # … with 113 more rows