Browse available RBA data series

Use these functions to find the table number or series ID of the data you're interested in.

browse_rba_series(search_string = "")

browse_rba_tables(search_string = "", refresh = FALSE)

rba_list_tables(...)

Arguments

search_string
Word or phrase to search for, such as "gold" or "commodity" or "labour". If left as "", the function will return all series.

refresh
logical; FALSE by default. When FALSE, internal data is used. When TRUE, the RBA website is re-scraped to obtain current information about available tables This can take a few seconds.

...
arguments to rba_list_tables() passed to browse_rba_tables()

Value

A data.frame (tbl_df) containing RBA data series/tables that match the search_string. Where no search_string is supplied, the data.frame will contain information about all RBA series/tables.

The data.frame returned by browse_rba_tables() includes a column called readable. This column takes the value TRUE if the table is able to be read by read_rba() and FALSE if it cannot be read.

Details

rba_list_tables() is a wrapper around browse_rba_tables() and is provided for compatibility with a previous package.

Examples

# Find series that contain 'unemployment'
browse_rba_series("unemployment")

#> # A tibble: 2 x 8
#>   table_no series series_id series_type table_title cur_hist description
#>   <chr>    <chr>  <chr>     <chr>       <chr>       <chr>    <chr>      
#> 1 H5       Unemp… GLFSUPSA  Seasonally… H5 Labour … current  Unemployed…
#> 2 H5       Unemp… GLFSURSA  Seasonally… H5 Labour … current  Unemployed…
#> # … with 1 more variable: frequency <chr>

# Or all labour-related series
browse_rba_series("labour")

#> # A tibble: 23 x 8
#>    table_no series series_id series_type table_title cur_hist description
#>    <chr>    <chr>  <chr>     <chr>       <chr>       <chr>    <chr>      
#>  1 H4       Non-f… GNFPROSQI Seasonally… H4 Labour … current  Non-farm l…
#>  2 H4       Quart… GWPIPRIQP Seasonally… H4 Labour … current  Wage price…
#>  3 H4       Quart… GWPIPUBQP Seasonally… H4 Labour … current  Wage price…
#>  4 H4       Quart… GWPIQP    Seasonally… H4 Labour … current  Wage price…
#>  5 H4       Year-… GNFAEYP   Seasonally… H4 Labour … current  Non-farm a…
#>  6 H4       Year-… GNFPROSQP Seasonally… H4 Labour … current  Non-farm l…
#>  7 H4       Year-… GNFULCYP  Seasonally… H4 Labour … current  Non-farm n…
#>  8 H4       Year-… GWPIPRIYP Seasonally… H4 Labour … current  Wage price…
#>  9 H4       Year-… GWPIPUBYP Seasonally… H4 Labour … current  Wage price…
#> 10 H4       Year-… GWPIYP    Seasonally… H4 Labour … current  Wage price…
#> # … with 13 more rows, and 1 more variable: frequency <chr>

# Or those related to commodities
browse_rba_series("commodities")

#> # A tibble: 9 x 8
#>   table_no series series_id series_type table_title cur_hist description
#>   <chr>    <chr>  <chr>     <chr>       <chr>       <chr>    <chr>      
#> 1 I2       Bulk … GRCPBCAD  Original    I2 Commodi… current  Index of c…
#> 2 I2       Bulk … GRCPBCSDR Original    I2 Commodi… current  Index of c…
#> 3 I2       Bulk … GRCPBCUSD Original    I2 Commodi… current  Index of c…
#> 4 I2       Bulk … GRCPBCSAD Original    I2 Commodi… current  Index of c…
#> 5 I2       Bulk … GRCPBCSS… Original    I2 Commodi… current  Index of c…
#> 6 I2       Bulk … GRCPBCSU… Original    I2 Commodi… current  Index of c…
#> 7 I2       Commo… GRCPAISAD Original    I2 Commodi… current  Index of c…
#> 8 I2       Commo… GRCPAISS… Original    I2 Commodi… current  Index of c…
#> 9 I2       Commo… GRCPAISU… Original    I2 Commodi… current  Index of c…
#> # … with 1 more variable: frequency <chr>

# Or all series
browse_rba_series()

#> # A tibble: 4,382 x 8
#>    table_no series series_id series_type table_title cur_hist description
#>    <chr>    <chr>  <chr>     <chr>       <chr>       <chr>    <chr>      
#>  1 A1       Austr… ARBAAASTW Original    A1 Reserve… current  Australian…
#>  2 A1       Capit… ARBALCRFW Original    A1 Reserve… current  Capital an…
#>  3 A1       Depos… ARBALDEPW Original    A1 Reserve… current  Deposits (…
#>  4 A1       Excha… ARBALESBW Original    A1 Reserve… current  Exchange s…
#>  5 A1       Gold … ARBAAGFXW Original    A1 Reserve… current  Gold and f…
#>  6 A1       Notes… ARBALNOIW Original    A1 Reserve… current  Notes on i…
#>  7 A1       Other… ARBAAOAW  Original    A1 Reserve… current  Other asse…
#>  8 A1       Other… ARBALOLW  Original    A1 Reserve… current  Other liab…
#>  9 A1       Total… ARBAATAW  Original    A1 Reserve… current  Total RBA …
#> 10 A1       Total… ARBALTLW  Original    A1 Reserve… current  Total RBA …
#> # … with 4,372 more rows, and 1 more variable: frequency <chr>

# Or just look for tables that contain the word 'labour'
browse_rba_tables("labour")

#> # A tibble: 2 x 5
#>   title             no    url                        current_or_histor… readable
#>   <chr>             <chr> <chr>                      <chr>              <lgl>   
#> 1 Labour Costs and… H4    https://rba.gov.au/statis… current            TRUE    
#> 2 Labour Force      H5    https://rba.gov.au/statis… current            TRUE    

# Or all tables
browse_rba_tables()

#> # A tibble: 123 x 5
#>    title                   no    url                  current_or_histo… readable
#>    <chr>                   <chr> <chr>                <chr>             <lgl>   
#>  1 Liabilities and Assets… A1    https://rba.gov.au/… current           TRUE    
#>  2 Liabilities and Assets… A1.1  https://rba.gov.au/… current           TRUE    
#>  3 Monetary Policy Changes A2    https://rba.gov.au/… current           TRUE    
#>  4 Monetary Policy Operat… A3    https://rba.gov.au/… current           TRUE    
#>  5 Holdings of Australian… A3.1  https://rba.gov.au/… current           TRUE    
#>  6 Foreign Exchange Trans… A4    https://rba.gov.au/… current           TRUE    
#>  7 Daily Foreign Exchange… A5    https://rba.gov.au/… current           TRUE    
#>  8 Banknotes on Issue by … A6    https://rba.gov.au/… current           TRUE    
#>  9 Detected Australian Co… A7    https://rba.gov.au/… current           TRUE    
#> 10 Assets of Financial In… B1    https://rba.gov.au/… current           TRUE    
#> # … with 113 more rows

# To re-scrape the RBA website to ensure you have up-to-date information
# about available tables:
if (FALSE) {
browse_rba_tables(refresh = TRUE)
}