Browse available RBA data series

Use these functions to find the table number or series ID of the data you're interested in.

browse_rba_series(search_string = "")

browse_rba_tables(search_string = "", refresh = FALSE)

Arguments

search_string

Word or phrase to search for, such as "gold" or "commodity" or "labour". If left as "", the function will return all series.

refresh

logical; FALSE by default. When FALSE, internal data is used. When TRUE, the RBA website is re-scraped to obtain current information about available tables This can take a few seconds.

Value

A data.frame (tbl_df) containing RBA data series/tables that match the search_string. Where no search_string is supplied, the data.frame will contain information about all RBA series/tables.

The data.frame returned by browse_rba_tables() includes a column called readable. This column takes the value TRUE if the table is able to be read by read_rba() and FALSE if it cannot be read.

Examples

# Find series that contain 'unemployment'
browse_rba_series("unemployment")
#> # A tibble: 2 × 8
#>   table_no series         series_id series_type table_title cur_hist description
#>   <chr>    <chr>          <chr>     <chr>       <chr>       <chr>    <chr>      
#> 1 H5       Unemployment   GLFSUPSA  Seasonally… H5 Labour … current  Unemployed…
#> 2 H5       Unemployment … GLFSURSA  Seasonally… H5 Labour … current  Unemployed…
#> # … with 1 more variable: frequency <chr>

# Or all labour-related series
browse_rba_series("labour")
#> # A tibble: 23 × 8
#>    table_no series        series_id series_type table_title cur_hist description
#>    <chr>    <chr>         <chr>     <chr>       <chr>       <chr>    <chr>      
#>  1 H4       Non-farm lab… GNFPROSQI Seasonally… H4 Labour … current  Non-farm l…
#>  2 H4       Quarterly pr… GWPIPRIQP Seasonally… H4 Labour … current  Wage price…
#>  3 H4       Quarterly pu… GWPIPUBQP Seasonally… H4 Labour … current  Wage price…
#>  4 H4       Quarterly wa… GWPIQP    Seasonally… H4 Labour … current  Wage price…
#>  5 H4       Year-ended n… GNFAEYP   Seasonally… H4 Labour … current  Non-farm a…
#>  6 H4       Year-ended n… GNFPROSQP Seasonally… H4 Labour … current  Non-farm l…
#>  7 H4       Year-ended n… GNFULCYP  Seasonally… H4 Labour … current  Non-farm n…
#>  8 H4       Year-ended p… GWPIPRIYP Seasonally… H4 Labour … current  Wage price…
#>  9 H4       Year-ended p… GWPIPUBYP Seasonally… H4 Labour … current  Wage price…
#> 10 H4       Year-ended w… GWPIYP    Seasonally… H4 Labour … current  Wage price…
#> # … with 13 more rows, and 1 more variable: frequency <chr>

# Or those related to commodities
browse_rba_series("commodities")
#> # A tibble: 9 × 8
#>   table_no series         series_id series_type table_title cur_hist description
#>   <chr>    <chr>          <chr>     <chr>       <chr>       <chr>    <chr>      
#> 1 I2       Bulk commodit… GRCPBCAD  Original    I2 Commodi… current  Index of c…
#> 2 I2       Bulk commodit… GRCPBCSDR Original    I2 Commodi… current  Index of c…
#> 3 I2       Bulk commodit… GRCPBCUSD Original    I2 Commodi… current  Index of c…
#> 4 I2       Bulk commodit… GRCPBCSAD Original    I2 Commodi… current  Index of c…
#> 5 I2       Bulk commodit… GRCPBCSS… Original    I2 Commodi… current  Index of c…
#> 6 I2       Bulk commodit… GRCPBCSU… Original    I2 Commodi… current  Index of c…
#> 7 I2       Commodity pri… GRCPAISAD Original    I2 Commodi… current  Index of c…
#> 8 I2       Commodity pri… GRCPAISS… Original    I2 Commodi… current  Index of c…
#> 9 I2       Commodity pri… GRCPAISU… Original    I2 Commodi… current  Index of c…
#> # … with 1 more variable: frequency <chr>

# Or all series
browse_rba_series()
#> # A tibble: 4,478 × 8
#>    table_no series        series_id series_type table_title cur_hist description
#>    <chr>    <chr>         <chr>     <chr>       <chr>       <chr>    <chr>      
#>  1 A1       Australian d… ARBAAASTW Original    A1 Reserve… current  Australian…
#>  2 A1       Capital and … ARBALCRFW Original    A1 Reserve… current  Capital an…
#>  3 A1       Deposits (ex… ARBALDEPW Original    A1 Reserve… current  Deposits (…
#>  4 A1       Exchange set… ARBALESBW Original    A1 Reserve… current  Exchange s…
#>  5 A1       Gold and for… ARBAAGFXW Original    A1 Reserve… current  Gold and f…
#>  6 A1       Notes on iss… ARBALNOIW Original    A1 Reserve… current  Notes on i…
#>  7 A1       Other assets… ARBAAOAW  Original    A1 Reserve… current  Other asse…
#>  8 A1       Other liabil… ARBALOLW  Original    A1 Reserve… current  Other liab…
#>  9 A1       Total assets  ARBAATAW  Original    A1 Reserve… current  Total RBA …
#> 10 A1       Total liabil… ARBALTLW  Original    A1 Reserve… current  Total RBA …
#> # … with 4,468 more rows, and 1 more variable: frequency <chr>

# Or just look for tables that contain the word 'labour'
browse_rba_tables("labour")
#> # A tibble: 2 × 5
#>   title                         no    url              current_or_hist… readable
#>   <chr>                         <chr> <chr>            <chr>            <lgl>   
#> 1 Labour Costs and Productivity H4    https://www.rba… current          TRUE    
#> 2 Labour Force                  H5    https://www.rba… current          TRUE    

# Or all tables
browse_rba_tables()
#> # A tibble: 123 × 5
#>    title                                   no    url   current_or_hist… readable
#>    <chr>                                   <chr> <chr> <chr>            <lgl>   
#>  1 Liabilities and Assets – Summary        A1    http… current          TRUE    
#>  2 Liabilities and Assets – Detailed       A1.1  http… current          TRUE    
#>  3 Monetary Policy Changes                 A2    http… current          TRUE    
#>  4 Monetary Policy Operations – Current    A3    http… current          TRUE    
#>  5 Holdings of Australian Government Secu… A3.1  http… current          TRUE    
#>  6 Securities Lending Repurchase and Swit… A3.2  http… current          TRUE    
#>  7 Foreign Exchange Transactions and Hold… A4    http… current          TRUE    
#>  8 Daily Foreign Exchange Market Interven… A5    http… current          TRUE    
#>  9 Banknotes on Issue by Denomination      A6    http… current          TRUE    
#> 10 Detected Australian Counterfeits by De… A7    http… current          TRUE    
#> # … with 113 more rows

# To re-scrape the RBA website to ensure you have up-to-date information
# about available tables:
if (FALSE) {
browse_rba_tables(refresh = TRUE)
}